[STYLE] `uint256` -> `uint`

+++ hardhat/contracts/ERC721FixedPricePurchase.sol
@@ -13,28 +13,28 @@ import "./interfaces/IOwnable.sol";
 /// @dev The collection owner fee is dependent on the existence of an `owner` function on the ERC721 contract
 contract ERC721FixedPricePurchase is Ownable {
     /// @dev mapping from ERC721 address -> token id -> listing price
-    mapping(address => mapping(uint256 => uint256)) public listing;
+    mapping(address => mapping(uint => uint)) public listing;
 
     /// @dev mapping from collection owner to collection fee
-    mapping(address => uint256) public collectionFee;
+    mapping(address => uint) public collectionFee;
 
     /// @dev mapping from collection owner to fees accrued
-    mapping(address => uint256) public collectionFeesAccrued;
+    mapping(address => uint) public collectionFeesAccrued;
 
     /// @dev fee for the protocol
-    uint256 public protocolFee;
+    uint public protocolFee;
 
     /// @dev protocol fees accrued
-    uint256 public protocolFeesAccrued;
+    uint public protocolFeesAccrued;
 
     /// @dev used to calculate the basis point fee
     uint constant FEE_BASE = 10000;
 
-    event Listed(address indexed erc721, uint256 indexed tokenId, address indexed owner, uint256 price);
+    event Listed(address indexed erc721, uint indexed tokenId, address indexed owner, uint price);
 
-    event Purchased(address indexed erc721, uint256 indexed tokenId, address indexed buyer);
+    event Purchased(address indexed erc721, uint indexed tokenId, address indexed buyer);
 
-    modifier onlyErc721Owner(address erc721, uint256 tokenId) {
+    modifier onlyErc721Owner(address erc721, uint tokenId) {
         require(IERC721(erc721).ownerOf(tokenId) == msg.sender, "ERC721FixedPricePurchase: Only ERC721 owner can call this function");
         _;
     }
@@ -50,7 +50,7 @@ contract ERC721FixedPricePurchase is Ownable {
     /// @param erc721 token contract
     /// @param tokenId id of the ERC721 token being listed
     /// @param price amount buyer must pay to purchase
-    function list(address erc721, uint256 tokenId, uint256 price) onlyErc721Owner(erc721, tokenId) public {
+    function list(address erc721, uint tokenId, uint price) onlyErc721Owner(erc721, tokenId) public {
         listing[erc721][tokenId] = price;
         emit Listed(erc721, tokenId, msg.sender, price);
     }
@@ -59,15 +59,15 @@ contract ERC721FixedPricePurchase is Ownable {
     /// @dev basis point fees are calculated using the fee base constant
     /// @param erc721 token contract
     /// @param tokenId id of the ERC721 token being listed
-    function purchase(address erc721, uint256 tokenId) public payable {
+    function purchase(address erc721, uint tokenId) public payable {
         require(msg.value >= listing[erc721][tokenId], "ERC721FixedPricePurchase: Buyer didn't send enough ether");
         require(listing[erc721][tokenId] > 0, "ERC721FixedPricePurchase: Token is not listed");
         listing[erc721][tokenId] = 0;
         address from = IERC721(erc721).ownerOf(tokenId);
 
-        uint256 collectionFeeAmount = msg.value * collectionFee[erc721] / FEE_BASE;
-        uint256 protocolFeeAmount = msg.value * protocolFee / FEE_BASE;
-        uint256 sellerFeeAmount = msg.value - protocolFeeAmount - collectionFeeAmount;
+        uint collectionFeeAmount = msg.value * collectionFee[erc721] / FEE_BASE;
+        uint protocolFeeAmount = msg.value * protocolFee / FEE_BASE;
+        uint sellerFeeAmount = msg.value - protocolFeeAmount - collectionFeeAmount;
 
         collectionFeesAccrued[erc721] += collectionFeeAmount;
         protocolFeesAccrued += protocolFeeAmount;
@@ -84,13 +84,13 @@ contract ERC721FixedPricePurchase is Ownable {
     /// @notice set the basis point fee of the collection owner
     /// @param erc721 token contract
     /// @param fee basis point amount of the transaction
-    function setCollectionFee(address erc721, uint256 fee) onlyCollectionOwner(erc721) public {
+    function setCollectionFee(address erc721, uint fee) onlyCollectionOwner(erc721) public {
         collectionFee[erc721] = fee;
     }
 
     /// @notice set the basis point fee of the protocol owner
     /// @param fee basis point amount of the transaction
-    function setProtocolFee(uint256 fee) onlyOwner() public {
+    function setProtocolFee(uint fee) onlyOwner() public {
         protocolFee = fee;
     }
 
@@ -99,7 +99,7 @@ contract ERC721FixedPricePurchase is Ownable {
     function collectionWithdraw(address erc721) public {
         require(collectionFeesAccrued[erc721] > 0, 'ERC721FixedPricePurchase: No funds to withdraw for this collection');
         address payable collectionOwner = payable(IOwnable(erc721).owner());
-        uint256 amount = collectionFeesAccrued[erc721];
+        uint amount = collectionFeesAccrued[erc721];
         collectionFeesAccrued[erc721] = 0;
         (bool sent,) = collectionOwner.call{value: amount}("");
         require(sent, "ERC721FixedPricePurchase: Failed to send Ether");
@@ -108,7 +108,7 @@ contract ERC721FixedPricePurchase is Ownable {
     /// @notice allows protocol owner to withdraw collected fees
     function protocolWithdraw() public {
         require(protocolFeesAccrued > 0, 'ERC721FixedPricePurchase: No protocol funds to withdraw');
-        uint256 amount = protocolFeesAccrued;
+        uint amount = protocolFeesAccrued;
         protocolFeesAccrued = 0;
         (bool sent,) = owner().call{value: amount}("");
         require(sent, "ERC721FixedPricePurchase: Failed to send Ether");
